Make photo class, initialize it and moving some methods to db objct and improve self to static
