+++ admin/includes/init.php
@@ -5,4 +5,5 @@ require_once("config.php");
 require_once("database.php");
 require_once("db_object.php");
 require_once("user.php");
+require_once("photo.php");
 require_once("session.php");
