+++ admin/includes/photo.php
@@ -1,0 +1,9 @@
+<?php
+
+
+class Photo extends DB_Object
+{
+
+}
+
+$photo = new Photo();
